[OPIK-6169][SDK] Refactor OpenTelemetry integration package structure.
- Moved `opik.otel` modules to `opik.integrations.otel` for better organization.
- Updated imports across tests and source code to reflect the new structure.
